[ch #159053998] Slit test and added utils for shared functions.
