+++ authors/apps/authentication/tests/utils.py
@@ -1,0 +1,16 @@
+""" utils provide data and methods used for testing. """
+from django.contrib.auth import get_user_model
+
+TEST_USER = {
+    "user": {
+        "email": "test@mail.com",
+        "password": "testpassword",
+        "username": "testuser"
+    }
+}
+
+def create_user(username="testuser", email="testuser@mail.com", password="password"):
+    """ create_user() creates and returns a user. """
+    user = get_user_model().objects.create_user(username, email, password=password)
+    user.save()
+    return user
